embeddings_try/requirements.txt
google-generativeai
numpy
dotenv